StoreSeed should not create a buffer
The typical use-case of this routine it to store an updated seed after
requesting multiple pseudo random numbers. This means that there is
really no need to be prepared for creating the seed buffer as it has
already been done. It is undesirable to create the seed buffer at this
point as it clobbers too many registers, there is really no suitable
place where a result can be kept. Now throw a NONEXISTENT error in
case the seed buffer does not exist.