Make SEED a secondary function
Normally we rely on the Time module for creating a seed. However, it
is sometimes desirable to use the same series of pseudo random
numbers, so it makes sense to have a SEED function. As it probably
will be used only once (if at all) in a program, it is a good
candidate for being a secondary function.