Add R/S function replacement
This function (named STOP') is needed to handle OS4 style pause
termination.

When pausing we stop on pressing R/S, but the standard R/S
function (named STOP) only knows about the system pause flag, not the
one internal to OS4.

It is not possible to have both flags on as the run loop goes through
mainframe and that would result in that the standard X display is
shown instead of any replacement application shell one, which is the
whole purpose of the Boost/OS4 PAUSE function.

Thus, we need to have an R/S replacement that is aware of the OS4
style pause and that can sense it in progress, terminate it in that
case and stop program execution.