Provide auto-assign top row keys disabling
Add LKAOFF and LKAON to disable and enable the top row
auto-assignments. This works as we have a system shell that is always
active so we can make use of a global flag in the system buffer to
disable it, rather than doing it the usual way and fiddle with the
assignment bits.