Accept removing non-existent buffers
Do not give an error if the buffer does not exist, simply accept
it. The idea is that you do not want the buffer and this makes it
easier for the user. If the user is interested in whether the buffer
actually exists, BUF? can be used.

If the buffer number is out of valid range a DATA ERROR is returned.